Moved ConcludeBattleDto to subpackage
 Changes to be committed:
	modified:   src/main/java/com/ardaslegends/presentation/api/BattleRestController.java
	renamed:    src/main/java/com/ardaslegends/service/dto/war/ConcludeBattleDto.java -> src/main/java/com/ardaslegends/service/dto/war/battle/ConcludeBattleDto.java
