+++ src/main/java/com/ardaslegends/service/dto/war/battle/ConcludeBattleDto.java
@@ -1,6 +1,7 @@
-package com.ardaslegends.service.dto.war;
+package com.ardaslegends.service.dto.war.battle;
 
 import com.ardaslegends.service.dto.player.DiscordIdDto;
+import com.ardaslegends.service.dto.war.SurvivingUnitsDto;
 
 public record ConcludeBattleDto(
         Long battleId,
